Fix issue where broker_notify_properties changes did not migrate.
Because the properties were defined as keys in a Hash, any removals made
by the user would not be properly migrated.  We only ever had them
defined as a Hash in order to be able to do efficient lookups.  Instead,
we can define them in the settings as a simple Array, and just convert
that Array to a Set to get the efficient lookups.  Arrays are properly
handled as expected by the config revamp, so this change leverages that.

Fixes #7677

(transferred from ManageIQ/manageiq@d6ec1eeff40b6703205560431ee3144967836462)