Empty VM Notes doesn't clear container notes field
When you clear the VM Notes field of a VM instead of setting
hardware[:annotation] to nil we leave it unset so it the value
in the database isn't cleared out.  To fix this set the
hardware[:annotation] field to nil explicitly if the
config.annotation property is blank.

https://bugzilla.redhat.com/show_bug.cgi?id=1326866


(transferred from ManageIQ/manageiq@2386d0ab5379a3b20cb1d7b3225032a45bb97efe)