Fix for VM linked to incorrect Storage
If there are multiple different Storages in the environment with
the same name, the vmLocalPathName will only point to one of them
due to a hash collision when using the datastore name as a key.

To fix this we use the MOR from the vm_inv['datastore'] property
instead of trying to find the storage_uid from the vmLocalPathName
when looking for the datastore the VM resides on.

https://bugzilla.redhat.com/show_bug.cgi?id=1334789


(transferred from ManageIQ/manageiq@ad2fc3bc7f767169aa2689e16144210cc5caf906)