Add orchestration status support for VMware Cloud
This patch adds the ability to get the current status of the
orchestration stack (in case of VCD, this is a vApp). The status is
retrieved from the EMS on demand. Corresponding tests are used to
validate the behaviour.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>


(transferred from ManageIQ/manageiq@9ec8e06774793880cdb7b1f6c4914515ba7b9e38)