Don't stop at a datacenter when filtering vm inv
When doing targeted refresh of a VM, if the VM's host isn't in inventory
then the folders above datacenter are not returned.  This causes
link_ems_inventory to set a new root folder for the EMS and clears all
other folder relationships.

This can happen if the user doesn't have permission to access the
cluster/host the VM is running on but does have access to the folder the
VM is in.

https://bugzilla.redhat.com/show_bug.cgi?id=1387336


(transferred from ManageIQ/manageiq@39fb6ce4c02e47030423da1abb7ee47b89e77f06)