Add VM specific customisations to vApp orchestration
Previously users were only allowed to modify the name of the provisioned
vApp and parameters that control the power state of the vApp.

With this change we are adding support for changing the name of the
instances that are deployed as part of the vApp as well as the ability
to select a different VDC network to which instances may be connected.
Similarly to the vCloud Director, each instance may be connected to a
different network.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>


(transferred from ManageIQ/manageiq@20b9aa30d8562b93ad60a142d77dfdd801e3058e)