Retrieve host storage devices host-by-host
To reduce memory usage retrieve host storage devices host by host
instead of in one call.  This was originally done to increase
performance but the vast majority of the overhead was due to running
garbage collection on every API call.  Now that we don't do this the
performance increase of retrieving all hosts in one call is greatly
reduced.

https://bugzilla.redhat.com/show_bug.cgi?id=1448963