Reverse the default of wait_for_worker_monitor
A majority of workers do not need to wait for the server to start, so
changing the default simplifies the code and also identifies which
workers actually need to wait.  In researching this, we found that the
only reason that workers are waiting is to actually wait for the vmware
broker to start and provide its connection pool.  Follow up PRs will
change this from wait_for_worker_monitor to "wait for broker", and we
can further make this more specific to vmware only classes.

See also
https://github.com/ManageIQ/manageiq/pull/14257#issuecomment-317818016
for a description of how the wait_for_worker_monitor manifests as a
"wait for broker" scenario.

(transferred from ManageIQ/manageiq@9e99ba02ebdda05cb072ee36a54622227bb1ac52)