Allow type of controller to be passed in disk opts
When adding a new disk when a new controller is needed we currently try
to use the same type of controller that is already on the VM or we have
a default that we use.  If a user wants to specify specifically what
type of controller they want there is no way to do that.  This adds the
option to the disk options hash to have the new scsi controller type
explicitly defined by the user.

https://bugzilla.redhat.com/show_bug.cgi?id=1428284