Inventory VM snapshots
With this commit we support VM snapshot inventoring. Since vCloud is
quite modes with information it reveals about VM snapshots (only creation time,
size and poweredOn, which tells if if the virtual machine was powered on when the
snapshot was created) we need to compose `uid` and `ems_ref` dynamically.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>
Signed-off-by: Sašo Cvitkovič <saso.cvitkovic@xlab.si>