Allow user to pick administrator password upon vApp provisioning
With this commit we add yet another two fields to the vApp provisioning
Service Dialog: admin_password (text input) and admin_reset (checkbox)
for each VM to allow for administrator password customization.

Passoword is displayed in plaintext, because this is also the case on vCloud
dashboard, so it makes no sense to hide it here.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>