Skip "none" vApp network when inventoring
With this commit we prevent NetworkManager from crashing is case when there
is a special network "none" listed in vApp network configuration section. This
"none" network is a reserved vApp network name which is for some reason present
in case when at least one NIC of the vApp VMs is not connected to any network.

Following XML is present for such "none" networks:

```
<vcloud:NetworkConfig networkName="none">
    <vcloud:Description>This is a special place-holder used for disconnected network interfaces.</vcloud:Description>
    <vcloud:Configuration>
        <vcloud:IpScopes>
            <vcloud:IpScope>
                <vcloud:IsInherited>false</vcloud:IsInherited>
                <vcloud:Gateway>196.254.254.254</vcloud:Gateway>
                <vcloud:Netmask>255.255.0.0</vcloud:Netmask>
                <vcloud:Dns1>196.254.254.254</vcloud:Dns1>
            </vcloud:IpScope>
        </vcloud:IpScopes>
        <vcloud:FenceMode>isolated</vcloud:FenceMode>
    </vcloud:Configuration>
    <vcloud:IsDeployed>false</vcloud:IsDeployed>
</vcloud:NetworkConfig>
```

and it must be ignored when inventoring.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>