Use utility function vm_powered_on? instead manual comparison
With this commit we replace manual comparison of status with
utility function. The logic is a bit more bullet-proof now since
we now only prevent VM delete when VM is "ON" while we originally
only allowed delete when VM was "OFF". So effectively we now allow
VM delete even if status is e.g. "unknown" which I think should succeed.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>