Completely stop/suspend VM, not just partially
If you stop/suspend VM in vCloud Director UI, it stops/suspends completely.
But when calling same action using vCloud API it only stops/suspends VM partially,
that is because it stays in `deployed` state. To reach the same effect as in
vCloud Director UI we now use`post_undeploy_vapp` with `UndeployPowerAction'`
set to `'powerOff'`/`'suspend'`

Signed-off-by: sasoc <saso.cvitkovic@gmail.com>