Shift fog-vcloud-director version to 0.1.10
With this commit we shift gem version because WebMKS console depends
on some API calls that were only implemented in this version of the
gem. Also, since the new gem shifts default version from 5.1 -> 5.5
and due to a bug that CloudProvider always uses default version no
matter what version is stored in VMDB - we update the spec to match
the new default.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>