Fixes ProvisionWorkflow#available_vlans_and_hosts
When the `super` form of this method returns `nil` for hosts, the
`.map(&:id)` call would no work properly.

In previous forms, `hosts` had `to_miq_a`, which will effectively run an
`Array.wrap` around the object.  I figured moving away from specialty
code for this was a better option moving forward, so I skipped that an
opted for an `unless` check instead.