Refactor relationships custom save block
Refactor the relationships custom save block to save from the top down
instead of bottom up.  This has a few benefits, namely there are
typically far fewer parents so the number of calls to add_child are
greatly reduced, and every call to add_child clears the relationships
cache so calling this for every child instead of every parent has a huge
impact.

On a 10,240 VM VC simulator this reduced the time for the
vm_parent_blue_folder_parents from 7min 47sec to 50 seconds.