Shift fog-vcloud-director version to 0.3.0
With this commit we shift fog-vcloud-director version to 0.3.0 which
is fog-core 2.1.x compliant, but remains fog-core 1.45 compatible.

Also, we omit the explicit fog-core requirement in vmware's Gemfile
bacause we don't need it per se so we better let the fog-vcloud-director
install whatever it needs.

Fixes https://github.com/ManageIQ/manageiq-providers-vmware/issues/319

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>