Don't require cloud tenant upon vApp instantiation
It used to be valid if user didn't pick anything in the Tenant field
on the vApp provisioning dialog - but now the form won't let you
through. Which is a problem for vCloud's vApp provisioning as we don't
even inventory Cloud Tenants, so the drop-down options are always
empty.

Looking at the code, the 'tenant_name' field is in fact marked as
required in the parent class of the vCloud's OrchestrationTemplate!

With this commit we fix the vCloud's OrchestrationTemplate so that it
now fully customizes its buttons, not just extends from base class.
Having such power in hand we perform following changes:

- drop `'tenant_name'` field entirely from our dialog
- copy `'stack_name'` field from the base class to our dialog
- modify `'stack_name'` field so that it now allows any string as vapp name
  because vCloud does not limit it in any way (even '!! 123 ?? ŠČŽ' is valid)

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1631216

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>