Use Optimist
Trollop is a deprecated gem, and has been renamed/replaced by Optimist.

This relies on https://github.com/ManageIQ/manageiq/pull/18246 being
merged for this to work properly.  Since this is a tool, however, this
shouldn't be a problem for the general use of the appliance (workers,
miq_server, etc.).