For Optimism (aka:  "Glass half full commit")
`rbvmomi` is our only external dependency that uses `optimist`, and it
just updated to use it in it's new version, which we are now a version
behind.

Required because of https://github.com/vmware/rbvmomi/pull/145 and our
recent switch to use `optimist` ourselves.

Thankfully, this change is less scary than it looks:

https://github.com/vmware/rbvmomi/compare/v1.13.0...v2.0.0