Check the ems.last_inventory_date before returning
It is possible that EmsRefresh can have a really early view of the VM
without all final data populated by the time the CloneVM_Task completes
which can lead to post-provision tasks failing if they depend on some
data in the vm record.

Waiting until the Ems' inventory date is at least as recent as the
CloneVM_Task completeTime we guarantee that the VM record represents the
final VM state.