Move the vmware refresh tests to a more standard place
The common location for refresh tests for the providers is in the
refresher_spec.  The streaming refresh tests were added to the
collector_spec so as not to collide with the main refresh specs but now
that legacy refresh has been deleted and streaming is the default we
should move these.