Use the portgroup key not ref as uid_ems
The portgroup config.key property is what is used in the VM NIC backings
to reference what portgroup they are on, not the portgroup's
MOID/ems_ref.

This happens typically when switches and portgrops are imported from
another vcenter.

Using the portgroup.key rather than the MOID allows us to link VM's
guest_devices up the their Lans properly.