Run post_refresh after save_inventory completes
There are some operations which are run after the refresh completes such
as connecting newly saved vms to ems_events that were caught prior to
the VM being saved.

After moving to streaming refresh only and the refresh no longer being
run in the context of the Refresher this functionality was lost.