Set :hidden on ems_clusters for "singular" clusters
If a host isn't in a "real" cluster (wsdl_class ClusterComputeResource)
e.g. it is shown directly under a datacenter or it is a standalone host
(no vCenter) then VMware has the host in a "singular" or "fake" cluster
where there can only be a single host child (wsdl_class
ComputeResource).

We can simplify the parent/child logic by relying on all hosts being in
an ems_cluster and only showing the ones that aren't "hidden" on the UI
similar to what is done for ems_folders.