Don't pass kind="leave" when deleting a dvportgroup
When a distributed virtual portgroup is deleted the objectUpdate has
kind="leave".  We have to parse portgroup updates at the whole switch
level due to our Switch -> Lan modeling but we have to change the
objectUpdate kind otherwise the switch parser thinks it is the switch
that is being deleted.