Fetch the ems_ref of a VMware Network
We model Lans at the host level not at the EMS level.  This results in
duplication as well as the lack of an ems_ref.

It is possible to get the ems_ref from the VMware Network object by
looking up a Network by name in the same datacenter as the current host.