Allow setting cpu/memory hot add/remove during provision
These settings control if the VM is able to have cpus added/removed
and memory added while the vm is running.

These can only be reconfigured while the vm is powered down so setting
these during provisioning removes the need for downtime in the future if
this has to be enabled.