Allow collection of disconnected hosts in inventory
Currently we skip any hosts that have a null `config` property, but also
have logic to save the connectionState of the host.  When a host is
disconnected the vSphere can't get the current `config` but will have
`summary.config`.  This means we are skipping all disconnected hosts
rather than reporting that they are disconnected.