Defer metrics collection batching to queue
Before:

Core partitions the metrics collection work and putting them onto the queue.
Vmware then double checks the work, adding partitioning.
This is unnecessary and in the process, overriding the concurrent_requests value
confuses core when it slices up the metrics collector requests.

After:

Vmware is now expecting the queue messages to be pre-partitioned. So no extra work is done here.