Fix VM Provision datastore location class
When building the `VirtualMachineRelocateSpec` the `datastore` property
is a `ManagedObjectReference` to a `Datastore`.  We were setting the
value to a simple `String` rather than a `VimString` with the proper
`xsiType`/`vimType`.

This would have broken when we moved from having VimString in ems_ref as
a serialized yaml column and added ems_ref_obj.